+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.6]
+### Added
+- Add `chatSession.isParticipantDisconnected()` method to check if participant has been disconnected from chat
+### Changed
+- Bump dependencies to patch security vulnerabilities
+
 ## [3.1.5]
 ### Changed
 - Remove verbose websocket incoming message logging from LpcConnectionHelper